Use x/net/context to support go<1.7.
