+++ lib/go/thrift/common_test_pre_go17.go
@@ -23,12 +23,10 @@ package thrift
 
 import "golang.org/x/net/context"
 
-// A processor is a generic object which operates upon an input stream and
-// writes to some output stream.
-type TProcessor interface {
-	Process(ctx context.Context, in, out TProtocol) (bool, TException)
+type mockProcessor struct {
+	ProcessFunc func(in, out TProtocol) (bool, TException)
 }
 
-type TProcessorFunction interface {
-	Process(ctx context.Context, seqId int32, in, out TProtocol) (bool, TException)
+func (m *mockProcessor) Process(ctx context.Context, in, out TProtocol) (bool, TException) {
+	return m.ProcessFunc(in, out)
 }
