+++ lib/go/thrift/pre_go17.go
@@ -23,12 +23,4 @@ package thrift
 
 import "golang.org/x/net/context"
 
-// A processor is a generic object which operates upon an input stream and
-// writes to some output stream.
-type TProcessor interface {
-	Process(ctx context.Context, in, out TProtocol) (bool, TException)
-}
-
-type TProcessorFunction interface {
-	Process(ctx context.Context, seqId int32, in, out TProtocol) (bool, TException)
-}
+var defaultCtx = context.Background()
