+++ lib/go/thrift/processor_go17.go
@@ -1,4 +1,4 @@
-// +build !go1.7
+// +build go1.7
 
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
@@ -21,7 +21,7 @@
 
 package thrift
 
-import "golang.org/x/net/context"
+import "context"
 
 // A processor is a generic object which operates upon an input stream and
 // writes to some output stream.
